+++ apps/client/src/environments/patch-notes.ts
@@ -1,6 +1,18 @@
 export const patchNotes = `### Bug Fixes
 
-* **db:** fixed a bug preventing collectability markers from showing on G2 skybuilders items.
-* **gearset:** fixed an issue with filters getting too few items for levels < 60.
-* **list:** fixed "mark as done in log" button for precrafts.
-* **workshop:** fixed permissions propagation.`;
+* **data:** fixed wrong aetheryte link for most npcs in cities.
+* **desktop:** fixed overlay slider not displaying properly.
+* **desktop:** fixed shortcut creation and firewall entries management.
+* **desktop:** fixed updater spawning error on application patching.
+* **list:** fixed permissions popup when opened from list panel.
+* **log-tracker:** log tracking is now tied to character, not account.
+* **profile:** fixed layout issue for non-verified characters.
+* **search:** disabled history if debounce is disabled while waiting for fix from antd.
+* **web:** removed inventory optimizer from menu in web app.
+
+
+### Features
+
+* **core:** zh and ko files are now loaded only if needed.
+* **desktop:** added \`-pid\` option flag to specify pid to attach to for packet capture.
+* **list:** removed performance mode and added automatic virtual scrolling.`;